+++ tests/testthat/test_L1000.R
@@ -19,6 +19,8 @@ test_that("L1000 metadata and conditions are loaded", {
     expect_is(conditions, "list")
     expect_identical(names(conditions), c("Perturbation type", "Cell line",
                                           "Dosage", "Time points"))
+
+    expect_true(!is.null(attr(l1000perturbationsSmallMolecules, "metadata")))
 })
 
 test_that("Perturbation types are retrievable", {
@@ -32,8 +34,24 @@ test_that("Compare using Spearman correlation", {
                                 method="spearman")
     expect_is(data, "l1000comparison")
     expect_identical(colnames(data), c(
-        "genes", "HepG2_spearman_coef", "HepG2_spearman_pvalue",
+        "compounds", "HepG2_spearman_coef", "HepG2_spearman_pvalue",
         "HepG2_spearman_qvalue", "Average_spearman_coef"))
+
+    expect_equal(head(data$compounds),
+                 c("CVD001_HEPG2_24H:BRD-A14014306-001-01-1:4.1",
+                   "CVD001_HEPG2_24H:BRD-A65142661-034-01-8:5.35",
+                   "CVD001_HEPG2_24H:BRD-K31030218-001-01-1:4.25",
+                   "CVD001_HEPG2_24H:BRD-K41172353-001-01-4:4.7",
+                   "CVD001_HEPG2_24H:BRD-K60476892-001-02-1:4.1072",
+                   "CVD001_HEPG2_24H:BRD-K62810658-001-05-6:4.6768"))
+
+    expect_equal(head(data[order(data$HepG2_spearman_coef), ]$compounds),
+                 c("CVD001_HEPG2_24H:BRD-A65142661-034-01-8:5.35",
+                   "CVD001_HEPG2_24H:BRD-K31030218-001-01-1:4.25",
+                   "CVD001_HEPG2_24H:BRD-K94818765-001-01-0:4.8",
+                   "CVD001_HEPG2_24H:BRD-K77508012-001-01-9:6.025",
+                   "CVD001_HEPG2_24H:BRD-K60476892-001-02-1:4.1072",
+                   "CVD001_HEPG2_24H:BRD-K84389640-001-01-5:4.225"))
 })
 
 test_that("Compare using Pearson correlation", {
@@ -41,20 +59,46 @@ test_that("Compare using Pearson correlation", {
                                 method="pearson")
     expect_is(data, "l1000comparison")
     expect_identical(colnames(data), c(
-        "genes", "HepG2_pearson_coef", "HepG2_pearson_pvalue",
+        "compounds", "HepG2_pearson_coef", "HepG2_pearson_pvalue",
         "HepG2_pearson_qvalue", "Average_pearson_coef"))
+
+    expect_equal(head(data$compounds),
+                 c("CVD001_HEPG2_24H:BRD-A14014306-001-01-1:4.1",
+                   "CVD001_HEPG2_24H:BRD-A65142661-034-01-8:5.35",
+                   "CVD001_HEPG2_24H:BRD-K31030218-001-01-1:4.25",
+                   "CVD001_HEPG2_24H:BRD-K41172353-001-01-4:4.7",
+                   "CVD001_HEPG2_24H:BRD-K60476892-001-02-1:4.1072",
+                   "CVD001_HEPG2_24H:BRD-K62810658-001-05-6:4.6768"))
+
+    expect_equal(head(data[order(data$HepG2_pearson_coef), ]$compounds),
+                 c("CVD001_HEPG2_24H:BRD-A65142661-034-01-8:5.35",
+                   "CVD001_HEPG2_24H:BRD-K31030218-001-01-1:4.25",
+                   "CVD001_HEPG2_24H:BRD-K94818765-001-01-0:4.8",
+                   "CVD001_HEPG2_24H:BRD-K77508012-001-01-9:6.025",
+                   "CVD001_HEPG2_24H:BRD-K84389640-001-01-5:4.225",
+                   "CVD001_HEPG2_24H:BRD-K60476892-001-02-1:4.1072"))
 })
 
 test_that("Compare using GSEA", {
     data <- compareAgainstL1000(diffExprStat, perturbations, cellLine,
                                 method="gsea")
     expect_is(data, "l1000comparison")
-    expect_identical(colnames(data), c("genes", "HepG2_WTCS", "Average_WTCS"))
-    expect_identical(head(data$genes),
-                     c("BRD-A14014306", "BRD-A65142661", "BRD-K31030218",
-                       "BRD-K41172353", "BRD-K77508012", "BRD-K84389640"))
-    expect_identical(head(data[order(data$HepG2_WTCS), ]$genes),
-                     c("BRD-A65142661", "caffeic-acid-phenethyl-ester",
-                       "BRD-K94818765", "BRD-K77508012", "BRD-K31030218",
-                       "BRD-K41172353"))
+    expect_identical(colnames(data),
+                     c("compounds", "HepG2_WTCS", "Average_WTCS"))
+
+    expect_equal(head(data$compounds),
+                 c("CVD001_HEPG2_24H:BRD-A14014306-001-01-1:4.1",
+                   "CVD001_HEPG2_24H:BRD-A65142661-034-01-8:5.35",
+                   "CVD001_HEPG2_24H:BRD-K31030218-001-01-1:4.25",
+                   "CVD001_HEPG2_24H:BRD-K41172353-001-01-4:4.7",
+                   "CVD001_HEPG2_24H:BRD-K60476892-001-02-1:4.1072",
+                   "CVD001_HEPG2_24H:BRD-K62810658-001-05-6:4.6768"))
+
+    expect_equal(head(data[order(data$HepG2_WTCS), ]$compounds),
+                 c("CVD001_HEPG2_24H:BRD-A65142661-034-01-8:5.35",
+                   "CVD001_HEPG2_24H:BRD-K96188950-001-04-5:4.3967",
+                   "CVD001_HEPG2_24H:BRD-K94818765-001-01-0:4.8",
+                   "CVD001_HEPG2_24H:BRD-K77508012-001-01-9:6.025",
+                   "CVD001_HEPG2_24H:BRD-K31030218-001-01-1:4.25",
+                   "CVD001_HEPG2_24H:BRD-K41172353-001-01-4:4.7"))
 })